(Chore) Clean up server side templating process
This commit re-factors the use of the express handlebars library into a wrapper file `template.js` that allows configuring helper methods only once. It also cleans up comments and left over console logs.
