+++ server/controllers/finance/reports/invoice.receipt.js
@@ -14,6 +14,7 @@
 var uuid        = require('node-uuid');
 var q           = require('q');
 var _           = require('lodash');
+var path        = require('path');
 
 var db          = require('../../../lib/db');
 var NotFound    = require('../../../lib/errors/NotFound');
@@ -37,7 +38,7 @@ const defaultRender = 'json';
 const FLAG_TRUE = 1;
 const SUCCESS_STATUS = 200;
 
-const template = './server/controllers/finance/reports/invoice.receipt.handlebars';
+const template = path.normalize('./server/controllers/finance/reports/invoice.receipt.handlebars');
 
 exports.build = build;
