+++ server/lib/renderers/json.js
@@ -1,17 +1,11 @@
 /**
- * @overview
- * BHIMA JSON Report Renderer (Wrapper)
- *
+ * @description
  * This server library is responsible for rendering reports in JSON format.
  * It adheres to the standard BHIMA Report Renderer API, accepting data
  * and options and returning a single compiled object.
  *
  * This renderer demonstrates the API however is mostly a wrapper to be used
  * uniformly with the other renderers (PDF/ HTML/ CSV).
- *
- * @todo    JSON Wrapper is currently just a wrapper to provide a uniform API, it
- *          could also provide functionality like JSON verification (JSON parse/
- *          JSON stringify).
  *
  * @module  lib/renderers/json
  */
@@ -29,6 +23,6 @@ exports.render = renderJSON;
  * @returns {Object}          JSON Object representing a report that can be sent
  *                            to the client.
  */
-function renderJSON(data, options) {
+function renderJSON(data, template, options) {
   return q.resolve(data);
 }
