+++ server/lib/renderers/pdf.js
@@ -1,39 +1,45 @@
-var wkhtmltopdf = require('wkhtmltopdf');
+/**
+ * @description 
+ * This service is responsible for producing PDFs on the server, it accepts handlebar templates and uses the html
+ * renderer to produce valid HTML - then streaming this through the wkhtmltopdf application to produce PDFs.
+ * 
+ * @requires wkhtmltopdf
+ * @requires stream-to-promise
+ * @requires path
+ * @requires q
+ */
+const wkhtmltopdf     = require('wkhtmltopdf');
+const q               = require('q');
+const streamToPromise = require('stream-to-promise');
+const process         = require('process');
+const path            = require('path');
 
-var html = require('./html');
+const html = require('./html');
 
-var path = require('path');
-var q = require('q');
-var streamToPromise = require('stream-to-promise');
-
-var process = require('process');
-
-var root = process.cwd();
-
-exports.render = renderPDF;
-
-function renderPDF(context, template, options) {
-
-  // pdf requires absolute path to be passed to templates to be picked up by wkhtmltopdf on windows
-  context.absolutePath = path.join(root, 'client');
-
-  console.log('using path', context.absolutePath);
-
-  return html.render(context, template)
-    .then(function (result) {
-      // var deferred = q.defer();
-
-      console.log('got html result success mayn', result);
-
-      /** @todo move to function */
-      // return streamToPromise(wkhtmltopdf(result, {}));
-      //
-      return streamToPromise(wkhtmltopdf(result));
-
-      // return deferred.promise;
-    });
-}
-
-exports.headers = {
+const headers = {
   'Content-Type' : 'application/pdf'
 };
+
+exports.render = renderPDF;
+exports.headers = headers;
+
+/**
+ * 
+ * @param {Object} context    Object of keys and values that will be made available to the handlebar template
+ * @param {String} template   Path to a handlebars template
+ * @returns {Promise}         Promise resolving in compiled PDF 
+ */
+function renderPDF(context, template) {
+  // pdf requires absolute path to be passed to templates to be picked up by wkhtmltopdf on windows
+  context.absolutePath = path.join(process.cwd(), 'client');
+
+  return html.render(context, template)
+    .then(function (htmlStringResult) {
+      // pass the compiled html string to the wkhtmltopdf process, this is just a wrapper for the CLI utility
+      let pdfStream = wkhtmltopdf(htmlStringResult);
+      
+      // this promise will only be resolved once the stream 'end' event is fired - with this implementation this will not allow the client to
+      // receive chunks of data as they are available
+      return streamToPromise(pdfStream);
+    });
+}
